doc/authors.texi is now built in the source directories
`doc/authors.texi' is a prerequisite of the Info output, so if it is
necessary to build it, the Info output (which is built in the source
directories by default) also needs to be rebuilt. If the user doesn't have
write access in the source directories, and this file needs to be generated
(configure has been rebuilt), then they will not be able to build the Info
outputs either. Therefore there was no point in building it in the build
directories.