NEWS file and webpage updated for version 0.1
Thew news file was updated in preparation for the first public release
(version 0.1). There were a lot of changes since the last update of the
NEWS file, but those changes are not a major concern the Gnuastro users at
this stage (that the libraries are not yet installable).

The webpage was also updated to include the links and remove the previous
notice (that the source code is not available).