Configure sets library flags automatically
Until now we were using Autoconf's `AC_CHECK_LIB' to check for the presence
of a library. In doing so, for its `ACTION-IF-FOUND' value, we were setting
an "absolutejunk" variable to intentionally avoid adding the library flags
for configuration time in the `LIBS' variable. This was based on my wrong
assumption at the time I first wrote these: I was thinking that linking
with a library that is not used by the program will be inefficient. I was
wrong, a library is searched, and only parts that are necessary are used in
the final executable, so if there is nothing to use in it, it won't be
used. The "absolutejunk" variable is therefore no longer used.

On the other hand, as recommended in the Autoconf manual, `AC_CHECK_LIB'
should be avoided in most common cases (we were not using it because of the
reason above). It will also allow us to accomodate other operating systems
if their installed libraries have another file name. So we are now using
`AC_SEARCH_LIBS' to check for the libraries. An great side-effect of
setting the `LIBS' variable, is that we no longer need the lower-level
library flags in checking for the higher-level libraries.

With this correction, the `-lLIBRARY' values of `AM_LDFLAGS' variable were
no longer necessary in any of the subdirectory makefiles. To make things
even more cleaner and avoid possible bugs due to repetition, the three
`CFLAGS', `LDFLAGS', and `CPPFLAGS' variables are now set at configure time
(while respecting Make's variable substitution of `top_builddir' and
`top_srcdir'). All of these similar definitions in each subdirectory are
now gone and can help us focus on the issues of that particular
subdirectory.

This fixes bug #48076.