Added link to tarball with bugfix in webpage
After fixing of bug #48076 in commit 51c907e (Configure sets library flags
automatically), a tarball was created and uploaded to the gnu.org
pre-release (alpha) webpage. So a link was added in the webpage explaining
the issue and with a link to the new tarball so affected users could
download.