Variable initialization reading FITS keywords
While testing compilation in Ubuntu (on a virtual machine), I noticed that
there was an uninitialized value warning for `valueptr' in the
`gal_fitsarray_read_keywords', so it was initialized to `NULL'. This would
not have been an issue since the `case' structure would go to `default' if
it wasn't set. But this will make it clear for the compiler, since
`valueptr' is not explictly checked.