Renamed fitsarrayvv library to fits
The fitsarrayvv library was one of the first libraries I wrote long before
Gnuastro. Its ancient relics are still available on Github (
https://github.com/makhlaghi/fitsarrayvv ). Actually, I wrote it even
before learning Git, so even the dates there can't show how old this name
is! Any way, at that time, I was only concerned with writing a wrapper on
CFITSIO to easily read and write a FITS array into a C array, so I chose
this very long name to only suite the job I had back then.

But it has since grown to do many more things: practically everything to do
with FITS files in Gnuastro. So the name was only just confusing and also
long. Now that we have the special `gnuastro/' directory to keep all the
headers, we can easily specify any generic name we like for the libraries
since they cannot be mixed with other headers. So I renamed this library to
`fits'. It can easily be included with `#include <gnuastro/fits.h>'.

With the change of the library name, all the functions which were defined
in it were also renamed. Correcting the function names, some of the
functions were also renamed to remove the redundant `fits' string in their
names. Ofcourse we still have to work on making them more readable and
clear for libraries, and this correction was a step towards that goal.

The last few functions were actually not related to FITS at all, they were
purely WCSLIB functions. So a new library (`wcs') was defined to keep these
types of functions. In the future we might want to include other file
formats (for example HDF5), so this modularization is necessary.