man pages built in doc/man
Previously the man page of all the utilities were made (using Help2man)
within that utility's particular directory. In order to build on multiple
threads (with `make -jN', N>1), it was then necessary to set the executable
as a dependency. But that would cause problems with cleaning later, because
the executable (a built target) was now a dependency of a distributed
object (the situation is fully explained in the "Errors with distclean"
section of the Automake manual). To avoid any errors in `make distclean',
we had to specify the man page as a clean target and thus if a user who was
building in the source directory would run `make clean' and they didn't
have Help2man they would loose the man pages (explained in bug #48124).

To fix the problem, building of the man pages of all the utilities is now
managed by `doc/Makefile.am'. After all, man pages are a form of
documentation. A `$(top_builddir)/doc/man' directory is built at configure
time to host the man pages. It already exists and is filled with the man
pages in the distributed tarball, so there is no problem for a user without
Help2man.

Previously the call to Help2man had a `-' at the start of the line. This
was recommended by Help2man (on its main webpage under "Using help2man with
make"). The reason was that we were using `AC_PATH_PROG' to check for the
existence of Help2man, so when it wasn't present, the value to the
`HELP2MAN' variable would become blank and the unreasonable command after
it would not cause an error in Make. The outcome of all of this were some
extra lines with an error in them that was ignored in all the `make'
outputs. This would make the output of `make' very dirty and could cause
confusion for the users. But now, there is no such problem and a user
building from the tarball will not see anything about building man pages or
help2man, or any ignored errors about it.

Besides a cleaner Make output, this process also significantly cleaned the
`src/*/Makefile.am' files and also removed one extra file from the tarball
sources for each utility (making it cleaner for an interested person to
read/inspect).

This fixes bug #48124.