Mandatory dependencies section edited/updated
In order to make things easier for someone who just wants to install the
dependencies from source, we used the "-latest.tar.gz" (or any alternative)
format instead of using "X.X.tar.gz" format. Also the tarball file name is
now a hyperlink in HTML and includes the address of the file in PDF. So
with a click a user can get it. They don't have to navigate through the
different webpage layouts to find what they want.

The name of the mandatory dependencies is also now a link to the main
webpage of those packages. While making these changes, the text was also
read over again and edited updated to be more clear.