Minor corrections to top webpage
A link to the info-gnuastro mailing list was added immediately under the
download link so interested people (who are not yet familiar with the
mailing lists!) can know about it. The "Download" section will be the most
visited part of the main webpage so having this link there will help guide
the users. A few other minor editorial corrections were made in the text.