Only necessary TeX Live packages explained in book
LaTeX (and some TeX packages) are necessary to bootstrap Gnuastro. In
"Bootstrapping dependencies" until now we had not made any particular
suggestions, so unexperienced enthusiasts would install the full ~4 giga
byte package. This can be a major problem for some interested people or
people who are not heavy users of TeX and could discourage people from
bootstrapping Gnuastor.

With the release of TeX Live 2016, I needed to update my own TeX Live, so I
thought of finding a way to bootstrap Gnuastro with the least amount of
downloaded and installed number of packages. After some tests, the "Basic
scheme" seemed to be the best installation scheme and the extra necessary
packages to successfully do the bootstrap were then found and included in
the book along with some explanations so a new commer can have their TeX
installation ready for the Gnuastro bootstrap easily.