Re-write and update of "Installation directory"
While this section was claiming to be useful for all applications when you
don't have root access, I found out that is not the case. In particular
when installing libraries with a `--prefix ~/.local' for example. So the
text was re-ordered in some places and generalized to also include library
installations for users that don't have root access and want to remain up
to date.