Matrix check in ImageWarp made more clear
The matrix check conditions in ImageWarp's 'ui.c' had an '&&' and '||'
called together. This created a warning in GCC 6.1.1: "warning: suggest
parentheses around ‘&&’ within ‘||’ [-Wparentheses]". The precedence of
'&&' is higher than '||' in C so the size conditions were not
ambiguous. However it would be more clear to have parenthesis.

To make the check more clear, the first condition (p->up.matrixname is
non-NULL) was moved to a separate 'if' statement and a parenthesis was put
around the first '&&'.