Clarification in the "Forking tutorial" section
The example branch name in the book's "Forking tutorial" section included
the bug ID. However, when the issue is a streightforward and simple fix, an
initial announcement is not necessary, therefore there is no bug-id yet. So
it is better (and also more descriptive) that the branch name be a very
short description of the issue and not a hard-to-read bug/task ID. So this
was corrected in the example. While doing this some of the text was also
edited to be more clear.