Make a symbolic link to access the tmpfs build
After the build was done in the tmpfs directory, it was hard to access it:
you had to 'cd /dev/shm/gnuastro' to get to it. So to make things easier,
the 'parallel-build-in-tmpfs' script will now create a symbolic link to
that location and put it in the top source directory so users can easily
get there by clicking on the link or running 'cd build'. This symbolic link
was also added to '.gitignore', so it is not tracked in the version
controlled history.