Removed argument on speed in parallel-build-in-tmpfs
Following some tests suggested by Mosè (in task #14100), we realized that
some bottle-neck is stopping the speed increase that we initially expected
from configuring and building in the RAM. So the speed argument is
currently removed from the explanations. Maybe later, if we find the
bottleneck, we can explain the reason so others are not as confused as we
were ;-)!

Still I believe the two advantages listed (having a clean source file and
not forcing two many read/writes on the HDDs and/or SSDs) are worth it to
keep the script for more testing.