Fix buffer overrun in src/imgcrop/imgcrop.c
The code logs a lines of predefined lenght, and the lenght is not too
long. So, a long file names overrun the buffer for a log line and make
the application crash. The fix truncs a long file names upto 30
characters before writing to the log.

This fixed the bugs 46241 and 45380