New macro for ImageCrop's verbose filename output
Besides the log-file, ImageCrop also prints the output file names on the
terminal in verbose mode. However, unlike the log-file, the terminal has a
limited width and to be readable, it is important for the verbose mode to
also manage the width of the things it prints. This had created bugs #46241
and #45380 which were fixed with the previous commit (cb9f299).

However, the buffer length until know was a hard-coded number (30) in the
code. So with this commit, that number is changed to a C Pre-Processor
macro which is configurable in 'src/imgcrop/main.h'.

Besides this macro a few other minor corrections were made:

  - The variable that keeps the filename length was changed to
    'outnamelen', from 'job_len'. The variable's properties have an 'out'
    in their names and this would make the code easier to read.

  - While the curly braces did follow the GNU style, their indentation and
    that of the lines between them did not. So this was also corrected.