All TABs in .c and .h files replaced with spaces
The coding convention that there should be no TABs in indentation was not
included in Gnuastro from the start. Therefore some of the code that was
written before its adoption still contained TAB characters used for
indentation. This was most evident in older utilities like ImageCrop
(historically, the first utility to be written within Gnuastro and some of
the older libraries. So with this commit, all the .c and .h files in 'lib/'
and '/src' were checked and any tabs were replaced with spaces.

Another common occurance of TAB characters was in multi-line macros (to
fill the space between the end of the code and the '/' character. These
occurances were also replaced with space characters.

This fixes bug #48692.