Tmpfs build directory name based on source dir name
In the script to configure and build Gnuastro in tmpfs, the name of the
temporary directory was 'gnuastro-XXXX'. This was contrary to the claim in
the comments that this file can be used in any program using the GNU Build
System. So now it uses 'basename' to find the name of the source directory
and use that to set the directory name in tmpfs.

With this system, when the script is run within a tarball, the temporary
directory will have two version numbers, maybe later we can work on
removing one of them. But for now this is sufficient, in practice a user
will hardly ever want to directly go to that directory, they will go
through the 'build' symbolic link and after a restart its gone with the
rest of the RAM.

When merging this commit with master, I forgot to include a notice that
this task is now complete. So it is being included here:

This finishes task #14100.