Tmpfs configure and build explanation moved inside book
A sub-sub-section was added to the book explaining the `tmpfs-config-make'
script for users. The explanation is actually taken from the comments at
the start of the script. So that long initial explanation was removed and
after a one-line explanation, the comments refer the curious user to the
proper part of the book to read more comfortably.