ImageCrop long filename segfault fix also in WCS mode
The bugs #45380 and #46241 reported a problem with long output filenames
causing segfaults in ImageCrop. The solution was found in commit
cb9f299. However, since the script creating the bug was only run in
ImageCrop's image mode, the fix was only applied to that mode. Exactly the
same problem also lay hidden in the WCS mode.

Therefore a function was defined for printing the outputs in verbose mode
and the fix was moved into that function. The function is called by both
image and WCS modes, so both modes can now benefit from that fix.