MakeProfiles also accepts WCS positions
Until now the profile centers in MakeProfiles could only be specified by
image coordinates (X and Y). However, for most practical purposes, the
World Coordinate System (WCS, using right ascension and declination) is
needed. So MakeProfiles was modified to also be able to read WCS positions
in the input catalog when the user asks so. To do this, two options were
added for the user to specify the RA and Dec columns. When these columns
are given, the X and Y coordinates are no longer used.

This process mainly involved the starting of MakeProfiles (mostly in
`ui.c'). The construction of the WCS structure (when a background image is
not provided) was moved to `ui.c' (from `mkprof.c'). Having the WCS
structure, the given RA and Decs are converted to X and Y, replaced with
the RA and Dec and the `xcol' and `ycol' values set to the old RA and Dec
columns. Reading the WCS structure in `ui.c' also simplified the output
writing parts of `mkprof.c'.

While doing this I noticed that if the given oversampling factor is not
odd, then it is incremented to become odd! This is very bad behavior (I
honestly don't know what I was thinking when I put this line!). If the user
is expecting to oversample by 2, we can't secretly oversample by 3 for
example, we will just greatly confuse the user! Instead, now an error
message is printed which explains that oversampling must be an odd number.

While doing the work, in some places, the indentation were also corrected
(mainly following the refactoring of function names in commit c63305c).

A simple test was also written to build profiles based on their input WCS
coordinates.

This finishes task #13566.