Script to rebuild for tests during development
During development, you often need to recompile the program and test your
work/changes in the code. With this script the process becomes really
easy. Especially when it is used in conjunction with the
`tmpfs-config-make' script of the top source directory. Just set the values
of the necessary variables in this script, then simply call it everytime
you want to re-build Gnuastro and test the changes you have made.

The comments at the start of the script thoroughly explain what it does. It
is just important for developers to remember that after their work is done
and they want to commit, they run first run the following command to reset
this file to its original status and not mistakenly commit their temporary
changes to this file.

   git checkout -- tests/during-dev.sh