MakeProfile's radeccat test input in tarball
In commit f5a1bfc, the ability to read the profile positions in the world
coordinate system was included and a test was written for it. However, I
forgot to run `make distcheck' before making the commit (and pushing to the
main repository). After running `make distcheck', I noticed that the test's
input catalog has not been included in the distribution files, so it will
be missing in tar-ball installs causing the test to fail. With this commit
the issue is solved.

While modifying the file, I also noticed that the `tests/during-dev.sh'
script (from commit cb7d44a) was mistakenly included in the distribution
tarball from the top-level `Makefile.am'. This script is now also managed
with the `tests/Makefile.am' now.

Also one part of the `tests/during-dev.sh' comments were made more clear
and corrected and a paragraph was added to the comments.