New macro to check columns for input catalogs
A new macro (`GAL_CHECKSET_CHECK_COL_NUM_IN_CAT') was added to
`lib/gnuastro/checkset.h' to only check the number of columns in a catalog
with a column that the user has specified. The old macro
(`GAL_CHECKSET_CHECK_COL_IN_CAT') would complain with an error on coming up
with such elements. The old macro was also modified to generally complain
for any non-finite number, not just NaNs.

This was mainly done as part of the work on task #14115 ("MakeProfiles flat
profiles have a given value"). For that task, it was necessary for the
catalog to have NaN values. Used when the user wants NaN covered profiles
to mask an elliptical region.