Macro to read numbers in txtarray is now a function
The CONVERTANDSAVE macro in `lib/txtarray.c' was changed to a new
`convertandsave' function. This makes debugging and checking of the results
much more easier. Also some comments were added for further clarification.

This was mainly done as part of the work on task #14115 ("MakeProfiles flat
profiles have a given value"), since the input catalog must contain NaN
values. Ultimately there was no problem here, but the code is now easier to
read and debug.