Flat profiles in MakeProfiles are given values
The values of the flat function profiles (codes 4 and 5) in MakeProfiles
were previously set using the two `--setconsttonan', `--setconsttomin'
options. It was thus not possible to make flat profiles, each with a
different value.

With this commit, those two MakeProfiles options were removed and a more
generic and much more cleaner mechanism was set to manage flat profiles. By
default, the flat profiles will have the total magnitude given in the
`--mcol' (magnitude column) of the catalog. However, if the new
`--mforflatpix' or `-F' option is used, the user can now use the same
magnitude column to specify the constant pixel value that s/he wants for
each specific profile. It is also possible to use NaN, or Infinity Floating
point values in this scheme.

The `--setconsttomin' process can now be replaced with a first call to
Arithmetic (which can give you the minimum value of the background image)
to get the minimum or maximum values of the background image, then using
that value in this column for all the profiles.

Internally, until now the individual profiles were made on a canvas of NaN
values. But that just made things so much more complicated. It would also
conflict with the generic and easy to follow process introduced in this
commit which would allow the creation of NaN profiles. Now, the plain
canvas is filled with values of `0.0f'. So it is not possible to set a
constant value of `0.0f' for the profiles. NaN (mask) profiles are much
more common than zero valued profiles. All the utilities in Gnuastro deal
NaN pixels as blank, but 0 valued pixels have meaning, so NaNs are much
better for a mask.

A small typo in `src/mkprof/ui.c' (the `printvalues' function) was also
corrected, `up->tcolset', was mistakenly written as `up->mcolset' (repeated
from the top two lines). Some other very minor issues were
corrected/cleared.

This finishes task #14115