MakeProfiles can now use input image as canvas
The `--inputascanvas' option was added to MakeProfiles to allow the users
to easily specify all the main canvas information from a reference
image. By canvas it is meant: the image size, the WCS information and the
blank (NaN) pixels.

This can make things like aperture photometry very easy, since the same
image that will be used in NoiseChisel (to get an estimate of the Sky value
and its standard deviation) can now be used to generate an image with
labeled regions to replace with NoiseChisel's object HDU. A test was also
added to check this on an image that has NaN pixels.

This finishes task #14116.