Key in NoiseChisel objects HDU, for clumps extension
NoiseChisel now adds a `WCLUMPS' (short for "with-clumps") header keyword
to its objects HDU. This header will later be used by MakeCatalog to see if
it should actually look for a clumps input and make a clumps catalog or
not. If it doesn't exit or be set to "no", then MakeCatalog will no longer
search for a clumps image or make a clumps catalog.

This can be very useful when clumps cannot be defined, for example
apperture photometry, where we only have specified regions (objects) and no
clumps.