Non-blank minimum and maximum for long in statistics
Two new functions were added to lib/statistics.c (and its header) for
finding the minimum and maximum values in a long array (while ignoring
blank values). All the functions until now were for float and double
types. Since NaN is not available for the long type, it was necessary to
check for the pixel not being a blank pixel.