Strings and non-existing keys in gal_fits_read_keywords
Until now if a key was not present in a HDU, `gal_fits_read_keywords' would
abort the program with an error. But the non-existence of a keyword might
be meaningful (the programmer might have alternatives, or use this
fact). So now, for each keyword that is read, the status is also stored in
the `gal_fits_read_header' structure and if it doesn't exist, there will be
no error.

Also the dealing of this function with strings was buggy due to the unclear
CFITSIO manual. So after looking into the codes of CFITSIO, I saw that it
expects the array to be set before hand and uses a fixed size of
`FLEN_VALUE' internally. So instead of a pointer, the
`gal_fits_header_keys' structure now has an array element.

In the process, some of the indentation issues that still remained from
before were corrected in these two files.