Type of output can be specified in MakeProfiles
MakeProfiles builds the profiles internally as a float array. Until now it
would also save the final output image in the same type unless the
background image had a different type. With the added capabilities to
specify the flat profile's pixel values, it is now necessary for the output
to have a certain data type (based on the flat values you wanted), for
example when making labels to be used in MakeCatalog, all the pixels have
integer values, and MakeCatalog also expects an integer type (int, short,
or long) for the labeled image.

With this commit, there is a new `--type' (`-T') option which allows the
user to set the output file type. Its values are given as a string of
standard types for easy typing on the commandline or in the configuration
files.

Given the separation of user-interface functions (like this) from the main
processing functions, implementing this new option is conceptually easy:
mainly to do with the `args.h' and `ui.c' files. The value given by the
user is immediately checked against the acceptable values and saved in the
`uiparams' structure as a CFITSIO BITPIX keyword value. If necessary it is
later copied to the `p->bitpix' variable which determines the type of the
output. With this more generic solution, the procedure to write the
function in the `write' function of `mkprof.c' also got much more simpler.

MakeProfile's `inputascanvas' test was also modified to save its value as a
long type to test this option.

Two other minor issues are corrected with this commit:

  - The replace option was only read from the command-line, not the
    configuration files.

  - When printing the value of the `--mforflatpix' option, it would
    mistakenly print the value to `--tcol'.