MakeCatalog makes clump catalog only when asked
MakeCatalogs will now read the `WCLUMPS' keyword in the header of the
objects HDU to see if it should make a clump catalog or not. If a
clump catalog isn't needed, then no input clump HDU will also be
necessary.

The value for `WCLUMPS' in the objects HDU should be a "yes" (not
case-sensitive). If it is anything else, or doesn't exist, then it is
assumed that no clump catalog is needed. To do the case insensitive check,
it was necessary to use the `strcasecmp' function which is not fully
portable in some systems, Gnulib's `strcase' module is now also imported
during the bootstrapping.

MakeCatalog (and more generally `gal_fits_read_keywords') can now deal
with the non-existence of a header keyword. So the procedures to read
other keywords were also updated: if they are not given, they are
calculated (where possible). To do this in an effective and clean
manner, a new `readkeywords' function was defined in MakeCatalog's
`ui.c'.

The detection and clump limiting S/N values (that can only be obtained
by NoiseChisel) are now given a NaN value if they aren't present in
the headers and will not be printed in the catalog's comments.