Aperture photometry test added for MakeCatalog
An aperture photometry test was added for the new feature of MakeCatalog
(to make a clumps catalog only when asked): `mkcatalog/aperturephot.sh'.

The input to the `mkprof/inputascanvas.sh' test was also changed to be the
same as NoiseChisel's input, so this new test could use the value of both
those tests. In doing so, the positions and values of the profiles (except
the truncation radius) was set to be similar to `mkprof/mkprofcat1.txt',
which was used to make the input image. Note that this required
oversampling so the `--oversample=1' option is now removed.

While doing the tests, some minor corrections were also made in
MakeCatalog:

  - The output name is now much more cleanly specified.
  - No clump image will be reported in the catalog comments.

This finishes task #14122.