Magnitude error column in MakeCatalog
A new column is now added to MakeCatalog for estimating the magnitude error
(called with the `--magnitudeerr' or `-e' option). This was done as part of
the work for task #14124. That task is not complete with this commit,
because we will also need to work on ways to incorporate correlated noise
and aperture error into this error too (and the S/N calculation in
general). The sky-region pseudo-detections and clumps can be very good
reference to make those estimations (maybe as a correction factor).