Template for creating a new utility added
All the utilities share the same basic source code and general options
structure. Until now, to add new utilities we would have to copy the files
of an existing utility and change all the relevant parts (the HEADER
utility was arguably the easiest).

To make things more clear and easy, a new `src/TEMPLATE' directory was
added which contains only the bare essentials of a utility. Any where that
we have to deal with the utility name, the word `TEMPLATE' is used. This
makes it really easy to start working on a new utility. It can also help
new developers see the bare minimum structure and start developing faster.