checkset func for some option value allocations
The `gal_checkset_allocate_copy_set' was defined to simplify the option
value allocation procedure and make the code more readable. It was
implemented in some utilities, but strangely it was not used in others. So
after a search through all `ui.c' and `args.h' files, it is now used in all
the places where it should have been used.