FITS table column information printed by Table
Table can now print the basic column information of a given FITS ASCII or
Binary table (so far only tested on a binary table). This is far from
complete for this utility, but for the time being it is a significant step
(it actually prints something useful!).

The type of the `gal_fits_read_hdu' function was changed from the exact
CFITSIO macro to an image or table binary value so we could use this
function in reading tables too.