Some code management in Table
To make Table more clear (easy to read) and easy to use, some modifications
were made (the function to print the column information is now in `ui.c'
and it is broken up into several functions for more general usage. One of
those new functions is now put in `lib/fits.c'.