Table accepts specified column number, name and regex
Using the `--column' option, users can now tell Table which columns of the
input table to read. The column can either be specified by its number, or
its name (or more generally a regex). We are using the GNU C Library's
implementation of regular expressions (that can also satisfy the POSIX
standard), so Gnulib's `regex' module is now also included as part of the
bootstrapping process.

While doing this, some work was also done on the linked list library:

  - An `x' was corrected in a function name of `linkedlist.h'. It was
    probably due to a miss-typed Emacs command that had gone unnoticed.

  - The `gal_linkedlist_sll_to_array' function can now make an array of
    size_t linked lists in reverse mode (same order as they were put in)
    with an added argument.