Blank CFITSIO pointers set by datatype, not bitpix
While working on the new Table utility, it was necessary to read a larger
collection of datatypes than those of the BITPIX keyword of FITS
images. CFITSIO defines its internal datatypes (starting with a `T') for
this purpose. Since the table columns and images have some types in common,
the more low-level parameter that should be used to set blank values and
allocate space should be the datatypes, not image or table type keywords
(which have different macros).

In this commit, the `gal_fits_bitpix_blank' function was replaced with
`gal_fits_datatype_blank' and now defines blank values for all CFITSIO
datatypes in its manual. It uses the GNU Scientific Library (GSL) complex
types for the complex datatypes.