Blank replacement function now accepts datatype
Following the discussion in commit 2071999 (Blank CFITSIO pointers set by
datatype, not bitpix), the old `blanktovalue' function now also accepts all
the CFITSIO dataypes. It is thus now named `gal_fits_blank_to_value' and
included in the header file incase some programs might need it. It is
currently not used anywhere, but can be useful so the effort that went to
into its initial definition is now kept for future use, especially after
libraries are installable.

While working on this, the following two issues were also addressed in this
commit:

  - The unsigned long datatype was missed in the commits since commit
    2071999, so it is now included in the functions dealing with type.

  - The CFITSIO specific Bitpix values are now also converted to datatype
    in `gal_fits_bitpix_to_dtype'.

  - Until now a single `GAL_FITS_FLOAT_BLANK' was used (assuming it is a
    value of NaN. But for some reason, it might be useful to have a
    different blank value for floats or doubles. Therefore a new
    `GAL_FITS_DOUBLE_BLANK' was also defined to be more general.