Configuration file in tests/during-dev.sh
`tests/during-dev.sh' now copies the configuration file (and makes a
`.gnuastro' directory to keep it in if necessary). During development the
configuration file might also be changed and updated and we don't want the
system-wide or user configuration files mixing up with our development, so
a `onlydirconf 1' line is also added to the copied configuration file.