Table prints contents of a FITS table
Table can now print the contents of a FITS table to the standard
output. There are several options to determine the formatting of the output
when it is printed. Currently for each column on each row a separate printf
statement is issued, the speed is not bad, but there might be more
efficient ways that we will have to work on.