FITS (binary) tables <--> plain text in Table
Table can now convert an ASCII (theoretically, not yet tested) and Binary
FITS table into an ASCII plain text file. It can also convert ASCII plain
text files to binary FITS tables. ASCII FITS tables have a different
`TFORMn' keyword format which made the work a little complicated so it was
not implemented at this time. Maybe later (when the need arises) we can
implement it is not hard, just needs some time (that I don't have now). Two
tests were also written to test these conversions.

Note that Table does have the `--fitstabletype' option to determine the
output FITS table type, but currently if `ascii' is passed to it, it will
inform the user that this is not yet implementend and that they should get
in touch with us so we can implement it.

Within Table, the old `tform' and `ttype' arrays in the `uiparams'
structure were named after the FITS standard keywords. However, these names
can be mis-leading, for example the "TTYPEn" is for the name of the column,
not the data type. So they were renamed to the more appropriate names:
`ttstr' (or table-type-string) and `tname'.

Table can still use a lot of work to greatly improve, in particular to read
plain text files by columns, this work will also benefit the `lib/txtarray'
(and thus all the utitlies that use it). Also once it is mature enough,
utilities that input or output tables/catalogs can do so from FITS
tables. So it would be a very useful work.

It would also be very interesting if we could adopt a human readable
convention to include the column information (for example name, type and
units) in a plain text file (for example as lines starting with `#c') so a
binary table and plain text file can be exactly convertable. Currently they
are not: the ascii to binary operation will read everything only as double.

The following modifications were made compared to previous commits:

 - `lib/checkset.c': A new `gal_checkset_check_file_report' function was
   added to report the existance of file, rather than just do the default
   operation. Some cosmetic changes were also made (fixing some line
   lengths).

 - `lib/fits.c': The `assert.h' header was an ancient artifact of the time
   I used the `assert' function! But I had forgot to remove it after
   abandoning that practice. So it was removed here. A new function
   `gal_fits_table_type' was added to return the type of a FITS table file.

 - `tests/during-dev.sh': a missed parenthesis was corrected.