Minor typo correction, Table task as complete
After merging the new Table utility in master, I forgot to mark the task as
finished in the commit message before pushing it to master. As a matter of
policy, we should not change any commit that has been pushed. So after
finding a small typo in the new section on Table, I am now marking the task
as complete in this commit.

This finishes task #13579.