Correctly add suffix when there wasn't any in input
The `gal_checkset_automatic_output' plays an important role in setting the
output files names when the user hasn't explicitly set an output name. If
the provided name didn't have a suffix, then this function would fail to
add a suffix to it.

The main cause of the problem was a `strcpy(out, inname)' call immediately
before the loop searching for an existing suffix. However, the loop was
assuming that the input and output strings are concatenated. In practice,
`strcpy' was redundant, so by removing it (and doing the search from the
end of the input string), we are able to achieve the same result while also
removing this bug. The situation is fully explained in the comments during
the loop.

This fixes bug #48899.