Output type print and checks moved to libraries
Until now only MakeProfiles had an option to determine the output data type
(BITPIX in FITS terminology). But Arithmetic (and soon probably other
utilities) can also greatly benefit from this and if it is common enough,
we can move the `--type' option to the common parameters.

In order to enable the easy usage of this option the following two new
library functions were defined: `gal_checkset_known_types' and
`gal_configfiles_print_type'. The first will make sure the given value is a
standard value and set the FITS macro for the type, the second will print
the string value for the given integer macro type when the user wants to
print the options and their values.

Both these functions were already defined in MakeProfiles, but are now
moved to the respective library and corrected in MakeProfiles.