Option for Arithmetic's output type
In commit ffd7f18 (Output type print and checks moved to libraries), the
functions to check the output image type (FITS BITPIX value) were moved to
the libraries to be accessible to all utilities. They were initially only
defined in MakeProfiles. In this commit we have used them in Arithmetic.

The type of Arithmetic's output is no longer an internal default, but a
configurable parameter by the user (through the `--type'
option. Internally, the used datatype is 'double', but this only concerns
the final creation of the output.