Static and shared libraries are now installed
As part of task #13765, the libraries are now installed along with the
utilities in the relevant system directories. The task greatly simplified
the Makefiles: all the utilities now only need one link option:
`-lgnuastro', instead of linking with all the separate libraries in one
option.

This does not yet complete `task #13765'. A new libraries chapter has to be
included in the book before we can mark that task as complete.